Implemented threading support
If a callback method is passed to the XBee constructor, a new thread will be spawned.
This thread will call the given callback method with any valid data received
from an XBee device. In order to clean up this thread properly, halt() must be called
on an XBee instance before closing its associated serial device.